Move functional tests to a separate module.
