+++ tests/test_parser.py
@@ -1,4 +1,3 @@
-from pathlib import Path
 import pytest
 import lkml
 import lkml.tokens as tokens
@@ -20,28 +19,6 @@ def parser():
         tokens.StreamEndToken(),
     ]
     return lkml.parser.Parser(stream)
-
-
-def load(filename):
-    """Helper method to load a LookML file from tests/resources and parse it."""
-    path = Path(__file__).parent / "resources" / filename
-    with path.open() as file:
-        return lkml.load(file)
-
-
-def test_block_with_single_quoted_field():
-    lookml = load("block_with_single_quoted_field.view.lkml")
-    assert lookml is not None
-
-
-def test_block_with_multiple_quoted_fields():
-    lookml = load("block_with_multiple_quoted_fields.view.lkml")
-    assert lookml is not None
-
-
-def test_block_with_nested_block():
-    lookml = load("block_with_multiple_quoted_fields.view.lkml")
-    assert lookml is not None
 
 
 def test_init_parser_index_starts_at_zero(parser):
